+++ src/commands/debugFile.ts
@@ -6,21 +6,22 @@ import { ExtensionContext } from '../extensionContext';
 /**
  * Debug the currently active Ruby editor.
  */
-export function debugEditor(context: ExtensionContext, textEditor: vscode.TextEditor): void {
+export async function debugEditor(
+  context: ExtensionContext,
+  textEditor: vscode.TextEditor,
+): Promise<boolean> {
   if (textEditor.document.languageId !== 'ruby') {
     vscode.window.showErrorMessage('Only Ruby files can be debugged');
-    return;
+    return false;
   }
+
   return debugFile(context, textEditor.document.uri);
 }
 
-/**
- * Debug the currently active Ruby editor.
- */
-export function debugFile(context: ExtensionContext, uri: vscode.Uri): void {
+async function debugFile(context: ExtensionContext, uri: vscode.Uri): Promise<boolean> {
   if (uri.scheme !== 'file') {
     vscode.window.showErrorMessage('Only local files can be debugged');
-    return;
+    return false;
   }
 
   const workspaceFolder = vscode.workspace.getWorkspaceFolder(uri);
@@ -33,5 +34,5 @@ export function debugFile(context: ExtensionContext, uri: vscode.Uri): void {
     runtimeExecutable: context.configuration.getRuntimeExecutable(workspaceFolder),
   };
 
-  vscode.debug.startDebugging(workspaceFolder, config);
+  return vscode.debug.startDebugging(workspaceFolder, config);
 }